+++ CHANGELOG.md
@@ -60,6 +60,12 @@
 
 
 
+### Snowpark Local Testing Updates
+
+#### New Features
+
+- Added support for NULL_IF parameter to csv reader
+
 ## 1.27.0 (2025-02-03)
 
 ### Snowpark Python API Updates